# api/openapi.yaml
openapi: 3.1.0
info:
  title: TX Carpool API
  version: 1.0.0
  description: |
    HTTP contract for the TX Carpool Telegram Mini App. User and admin endpoints
    use a Secure, HttpOnly session cookie established from validated Telegram
    Mini App `initData`. Provider callbacks use provider-specific signatures and
    are never authenticated by the browser session.

    All identifiers and timestamps are strings. TXB values use integer hundredths
    encoded as decimal strings. Traffic byte counts are decimal strings. The
    server is authoritative for prices, exchange rates, entitlement dates, and
    payment state.
  license:
    name: Proprietary
servers:
  - url: /
    description: Same-origin API
...
security:
  - cookieSession: []
paths:
...
  /api/v1/onboarding/username:
    put:
      tags: [Onboarding]
      operationId: reserveUsername
      summary: Reserve the immutable service username
      description: |
        Accepts only `^[a-z]{3,9}$`. The service checks local uniqueness, performs
        a Remnawave preflight, and also treats Remnawave `A019` as a duplicate if
        the upstream race is lost.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsernameRequest'
      responses:
        '200':
          description: Username reserved and onboarding advanced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
...
  /api/v1/subscription/revoke:
    post:
      tags: [Dashboard]
      operationId: revokeSubscription
      summary: Rotate the current Remnawave subscription URL
      description: Calls Remnawave revoke with `revokeOnlyPasswords:false`. The old URL must be treated as invalid after success.
      responses:
        '200':
          description: Rotated subscription reference.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionRotation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
...
  /api/v1/payments/orders:
    post:
      tags: [Billing]
      operationId: createPaymentOrder
      summary: Create a balance top-up order
      description: |
        The requested TXB minor units are converted using the current fixed-decimal
        dashboard rate. Provider payable amounts are rounded upward to provider
        precision. A successful authoritative callback credits exactly `txbMinor`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentOrderRequest'
      responses:
        '201':
          description: Durable payment order, QR payload, and provider navigation target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/UpstreamFailure'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: txc_session
      description: Opaque seven-day session created only by validated Telegram Mini App init data.
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing, invalid, or expired authentication or provider signature.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: The authenticated identity cannot perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Conflict:
      description: Request conflicts with current persisted state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UpstreamFailure:
      description: A required provider could not complete the synchronous portion of the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ID:
      type: string
      minLength: 1
      examples: [01JZX6BKW16M0CNQ5Q74K8DN7J]
    Timestamp:
      type: string
      format: date-time
      examples: ['2026-08-07T10:00:00Z']
    DecimalInteger:
      type: string
      pattern: '^-?\d+$'
      examples: ['1250']
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
          examples: [INSUFFICIENT_BALANCE]
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
        requestId:
          type: string
          minLength: 1
...
    OnboardingStep:
      type: string
      enum: [intro, membership, username, agreement, complete]
    User:
      type: object
      additionalProperties: false
      required:
        - id
        - telegramId
        - firstName
        - lastName
        - telegramUsername
        - username
        - role
        - onboardingState
        - groupJoined
        - channelJoined
        - policyAcceptedAt
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/ID'
        telegramId:
          type: string
          pattern: '^\d+$'
        firstName:
          type: string
        lastName:
          type: string
        telegramUsername:
          type: string
        username:
          type: [string, 'null']
          pattern: '^[a-z]{3,9}$'
        role:
          type: string
          enum: [user, admin]
        onboardingState:
          $ref: '#/components/schemas/OnboardingStep'
        groupJoined:
          type: boolean
        channelJoined:
          type: boolean
        policyAcceptedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
    AuthState:
      type: object
      additionalProperties: false
      required: [authenticated, user]
      properties:
        authenticated:
          type: boolean
          const: true
        user:
          $ref: '#/components/schemas/User'
...
    UsernameRequest:
      type: object
      additionalProperties: false
      required: [username]
      properties:
        username:
          type: string
          pattern: '^[a-z]{3,9}$'
...
    Money:
      type: object
      additionalProperties: false
      required: [currency, minor, display]
      description: Integer minor units. TXB, CNY, and USD use scale 2; XTR uses scale 0.
      properties:
        currency:
          type: string
          enum: [TXB, CNY, USD, XTR]
        minor:
          $ref: '#/components/schemas/DecimalInteger'
        display:
          type: string
          description: Server-formatted display value; never parse this for calculations.
...
    SubscriptionRotation:
      type: object
      additionalProperties: false
      required: [subscriptionUrl]
      description: The URL is a bearer credential and must not be logged, cached outside the authenticated UI, or exposed through admin list endpoints.
      properties:
        subscriptionUrl:
          type: string
          format: uri
...
    PaymentProvider:
      type: string
      enum: [ezpay, bepusdt, stars]
    PaymentStatus:
      type: string
      enum: [creating, pending, paid, expired, failed, refunded]
    PaymentOrderRequest:
      type: object
      additionalProperties: false
      required: [provider, txbMinor]
      properties:
        provider:
          $ref: '#/components/schemas/PaymentProvider'
        txbMinor:
          type: string
          pattern: '^[1-9]\d*$'
    PaymentOrder:
      type: object
      additionalProperties: false
      required:
        - id
        - provider
        - status
        - txb
        - payableAmount
        - payableCurrency
        - rateSnapshot
        - paymentUrl
        - qrPayload
        - expiresAt
        - paidAt
        - refundedAt
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: '#/components/schemas/ID'
        provider:
          $ref: '#/components/schemas/PaymentProvider'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        txb:
          $ref: '#/components/schemas/Money'
        payableAmount:
          type: string
          pattern: '^\d+(\.\d+)?$'
        payableCurrency:
          type: string
          enum: [CNY, USD, XTR, USDT]
        rateSnapshot:
          type: string
          description: Fixed-decimal rate used when the order was created.
        paymentUrl:
          type: [string, 'null']
          format: uri
        qrPayload:
          type: [string, 'null']
        expiresAt:
          $ref: '#/components/schemas/Timestamp'
        paidAt:
          type: [string, 'null']
          format: date-time
        refundedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
...